fix(api): address cli_events review — move schema to cli.ts, add created_at index, fix header
