+++ apps/api/drizzle/meta/_journal.json
@@ -50,6 +50,13 @@
       "when": 1779104230032,
       "tag": "0006_cynical_stone_men",
       "breakpoints": true
+    },
+    {
+      "idx": 7,
+      "version": "7",
+      "when": 1779104652269,
+      "tag": "0007_nervous_meggan",
+      "breakpoints": true
     }
   ]
 }
