+++ apps/api/src/db/schema/index.ts
@@ -1,5 +1,6 @@
 export * from "./auth";
 export * from "./challenge";
 export * from "./challenge-metadata";
+export * from "./cli";
 export * from "./email";
 export * from "./onboarding";
